fix: Refactor core to and remove input context
