+++ packages/graphback-core/src/annotations/Types.ts
@@ -1,0 +1,7 @@
+import { TypeDefinitionNode } from 'graphql';
+import { parseAnnotations } from 'graphql-metadata';
+
+export const parseTypeAnnotations = (node: TypeDefinitionNode) => {
+
+    return node.description ? parseAnnotations('crud', String(node.description)) : undefined;
+}
