--- packages/graphback-core/src/api/index.ts
@@ -1,5 +1,0 @@
-
-export * from "./ContextCreator"
-export * from "./ContextTypes"
-export * from "../plugin/GraphbackCRUDGeneratorConfig"
-export * from "./GraphbackOperationType"
