--- packages/graphback-core/src/graphql/index.ts
@@ -1,3 +1,0 @@
-
-export * from "./GraphQLContextCreator"
-export * from "./graphqlUtils"
