+++ packages/graphback-core/src/index.ts
@@ -1,9 +1,6 @@
 
 // Top level API
-export * from './api'
-export * from './graphql'
-export * from './logging';
-export * from './utils/fileUtils';
+
 export * from './utils/getRelationFieldName';
 export * from './utils/getBaseType';
 export * from './plugin/GraphbackPlugin'
