--- packages/graphback-core/src/logging/index.ts
@@ -1,1 +1,0 @@
-export * from './logger';
