--- packages/graphback-core/src/utils/lowerCaseFirstChar.ts
@@ -1,3 +1,0 @@
-export function lowerCaseFirstChar(text: string) {
-  return `${text.charAt(0).toLowerCase()}${text.slice(1)}`;
-}
